LD: optimizer fixes
1 and => nested if
2 unless => if .... #t ....
3 copy test-true/false "properties"

For 1 and 2, the problem is that the optimizer expects fully expanded
code but these are macros. Could we use the expander instead, though?

For 3, unsure if we can avoid the copying because we really do end up
with a new syntax object ... maybe the table (types/type-table ... if)
can use a different, weaker equality? Not sure. May also want to
syntax-track-origin when copying. What does the optimizer do?