LD double the sequence types for for loops
Fix transient for (for ((x (open-input-string "aaa"))) ....)
 by doubling the number of types for `make-sequence`

The transient error was "expected Byte got eof-object" from the
 pos->vals function returned by `make-sequence`

TR assumes a Byte and that's fine because there is a guard around
 the body of the for-loop. Why not make the type more accurate? The
 trouble is the guard is optional so we have:

 (if (if guard-fun (guard-fun next-val) #true)
   (.... for-body assuming next-val is good)
   (.... done with loop))

Occurrence typing can't handle this; we need a dependent type.

Thankfully case-> is strong enough, the only dependence is on #f or not.