LDblame: checkpoint stuck reviving types
May have solved the "revive" issue ... works for small example:

1. write '#,(type->sexp t) in all transient-asserts
2. revive with (eval datum ns) with suitable namespace

But hitting a bigger, different issue in transient-blame.rkt when trying
to revive in a real context. Issue happens without eval, using the type
constructors directly --- though not for the -Symbol type.

Problem is likely in my dynamic-require circularity ... what to do? ...
 can "everything" happen in an eval?