transient-blame: BROKEN attempt2 parsing types
Tried this:

- at compile-time, put a hash into the #%type-decl submodule
- at runtime, use hash to lookup types instead of calling eval

But it doesn't work, the current hash values give unbound identifier
errors!

This hash maps type-sexpressions to real types,
 done by calling type->sexp as it's intended --- to generate code that
 goes inside a type-decl submodule.

Maybe there'd be no problem if we could insert the types when we first
create the submodule ... but that's not possible we don't know all the
types that become contracts at that point ..... at least its not
convenient ... ok lets try one more thing.